requirements.txt
pycairo
pygame
svgwrite
board
RPi.GPIO
adafruit-circuitpython-seesaw
